Add local authentication
Use a local Passport strategy to authenticate users with the OGS API.

Users will log in with their OGS username and application password
(available in OGS settings). Passport will request an auth token from
OGS and redirect to home if successful. If the login is not successful,
the user will be redirected to the root path for now.